OMPL status after final tests at 3IT

+++ src/ompl/base/SafeMotionValidator.h
@@ -40,7 +40,7 @@ namespace ompl
 
             virtual bool checkMotion(const State *s1, const State *s2) const;
             virtual bool checkMotion(const State *s1, const State *s2, std::pair<State*, double> &lastValid) const;
-            bool checkMotion(const State *s1, const State *s2, int valid_segment_factor) const;
+            bool checkMotion(const State *s1, const State *s2, double valid_segment_factor) const;
 
             //STa test
             bool checkMotionTEST(const State *s1, const State *s2);
@@ -50,7 +50,7 @@ namespace ompl
             bool checkMotionIndividualLinksWithDist(const State *s1, const State *s2, double travel_dist_limit, double& min_obstacle_dist, bool fast_dist);
             bool checkMotionIndividualLinks(const State *s1, const State *s2, double travel_dist_limit, bool fast_dist);
             bool checkMotionSelfCCIndividualLinks(const State *s1, const State *s2, double travel_dist_limit) const;
-            bool checkMotionSelfCCDiscrete(const State *s1, const State *s2, int valid_segment_factor) const;
+            bool checkMotionSelfCCDiscrete(const State *s1, const State *s2, double valid_segment_factor) const;
             bool checkMotionSelfCCDiscreteWS(const State *s1, const State *s2, double travel_dist_limit) const;
 
             bool checkMotionWorldIndividualLinks(const State *s1, const State *s2, double travel_dist_limit, bool fast_dist) const;
@@ -59,8 +59,8 @@ namespace ompl
             double minObstacleDistMotionIndividualObjects(const State *s1, const State *s2, double travel_dist_limit, bool fast_dist, double& object_danger_factor) const;
             double minObstacleDistMotionIndividualObjectsTEST(const State *s1, const State *s2, double travel_dist_limit, bool fast_dist) const;
 
-            double minObstacleDistMotionDiscrete(const State *s1, const State *s2, int valid_segment_factor, bool fast_dist) const;
-            double minObstacleDistMotionDiscreteExact(const State *s1, const State *s2, int valid_segment_factor) const;
+            double minObstacleDistMotionDiscrete(const State *s1, const State *s2, double valid_segment_factor, bool fast_dist) const;
+            double minObstacleDistMotionDiscreteExact(const State *s1, const State *s2, double valid_segment_factor) const;
 
 
         private:
